chore(resharper): clean up warnings from codeql/12 style sweep
- UserHelper.cs: capture user into 'fallbackUser' local before
  GetOrCreate lambda so ReSharper AccessToModifiedClosure is silenced
- HttpHelper.cs: drop stale '<param name="memoryCache">' XML tag now
  that the other params don't have matching tags (was creating 5
  InvalidXmlDocComment warnings since the rename)
- test/ClinicalScheduler/{EmailNotificationTest,TestDataBuilder}.cs:
  replace '(double)(-7 * x)' with '-7.0 * x' — same intent (force the
  multiplication into double per cs/loss-of-precision) but no
  RedundantCast warning from ReSharper

+++ test/ClinicalScheduler/EmailNotificationTest.cs
@@ -119,8 +119,8 @@ namespace Viper.test.ClinicalScheduler
                 await _context.Weeks.AddAsync(new Week
                 {
                     WeekId = weekId,
-                    DateStart = DateTime.UtcNow.AddDays((double)(-7 * (10 - weekId))),
-                    DateEnd = DateTime.UtcNow.AddDays((double)(-7 * (10 - weekId) + 6)),
+                    DateStart = DateTime.UtcNow.AddDays(-7.0 * (10 - weekId)),
+                    DateEnd = DateTime.UtcNow.AddDays(-7.0 * (10 - weekId) + 6),
                     TermCode = 202501
                 });
             }
